Fix CE not processing alert status
Fixes #2968
